use snake_case instead of camelCase
In C, C++, and Python, we normally use snake_case to make the code
readable. snake_case rules:

- Write word using lower case English letters.
- Use underscore (_) to separate words if the variable name has more
  than 1 word.

Example of snake_case:
- line_number
- print_horizonal_line
- new_line

On the other hand, camelCase is usually used in language like Java, Go.
Example of camelCase:

- lineNumber
- printHorizonalLine
- newLine

It's OK to use camelCase, but in C, popular projects prefer to use
snake_case, so we follow them.